refactor: removed unnecessary circular reference
